Added Orbit Control for PlayerOne.js.
Added Orbit Control now when we click submit of three popup then we will move to orbit control page.

In that page we have a submit button which will take us to the Editor. And a back button on Editor Screen which will take us to the orbit control page.